Moving browser-opening/closing behavior out to page
