+++ robotpageobjects/base.py
@@ -8,6 +8,7 @@ import warnings
 from robot.utils import asserts
 from selenium import webdriver
 from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
 from selenium.common.exceptions import WebDriverException
 from selenium.webdriver.remote.webelement import WebElement
 from Selenium2Library import Selenium2Library
@@ -542,156 +543,6 @@ class _BaseActions(_S2LWrapper):
         self.set_selenium_timeout(self.selenium_implicit_wait)
 
         self.baseurl = self._option_handler.get("baseurl")
-        self.browser = self._option_handler.get("browser") or "phantomjs"
-        self.service_args = self._parse_service_args(self._option_handler.get("service_args", ""))
-
-        self._sauce_options = [
-            "sauce_username",
-            "sauce_apikey",
-            "sauce_platform",
-            "sauce_browserversion",
-            "sauce_device_orientation",
-            "sauce_screenresolution",
-        ]
-        for sauce_opt in self._sauce_options:
-            setattr(
-                self,
-                sauce_opt,
-                self._option_handler.get(sauce_opt)
-            )
-
-        self._attempt_sauce = self._validate_sauce_options()
-
-        # There's only a session ID when using a remote webdriver (Sauce, for example)
-        self.session_id = None
-
-    def _parse_service_args(self, service_args):
-        return [arg.strip() for arg in service_args.split(" ") if arg.strip() != ""]
-
-
-    def _validate_sauce_options(self):
-
-        # If any sauce options are set, at least
-        # username, apikey, and platform must be set, the rest are optional
-        sauce = {}
-        for attr in dir(self):
-            if attr.startswith("sauce_"):
-                sauce[attr] = getattr(self, attr)
-
-        at_least_one_sauce_opt_set = any(sauce.values())
-        if at_least_one_sauce_opt_set and (not sauce["sauce_username"] or
-                                               not sauce["sauce_apikey"] or not sauce["sauce_platform"]):
-            raise exceptions.MissingSauceOptionError("When running Sauce, need at " +
-                                                     "least sauce-username, sauce-apikey, and sauce-platform " +
-                                                     "options set.")
-
-        # If we get here, tell the object that it's going to
-        # attempt to use sauce and that all needed sauce options are
-        # at least set.
-        return at_least_one_sauce_opt_set
-
-    @not_keyword
-    def _resolve_url(self, *args):
-
-        """
-        Figures out the URL that a page object should open at.
-
-        Called by open().
-        """
-        pageobj_name = self.__class__.__name__
-
-        # determine type of uri attribute
-        if hasattr(self, 'uri'):
-            uri_type = 'template' if re.search('{.+}', self.uri) else 'plain'
-        else:
-            raise exceptions.UriResolutionError("Page object \"%s\" must have a \"uri\" attribute set." % pageobj_name)
-
-        # Don't allow absolute uri attribute.
-        if self._is_url_absolute(self.uri):
-            raise exceptions.UriResolutionError(
-                "Page object \"%s\" must not have an absolute \"uri\" attribute set. Use a relative URL "
-                "instead." % pageobj_name)
-
-        # We always need a baseurl set. This enforces parameterization of the
-        # domain under test.
-
-        if self.baseurl is None:
-            raise exceptions.UriResolutionError("To open page object, \"%s\" you must set a baseurl." % pageobj_name)
-
-        elif len(args) > 0 and hasattr(self, "uri") and self._is_url_absolute(self.uri):
-            # URI template variables are being passed in, so the page object encapsulates
-            # a page that follows some sort of URL pattern. Eg, /pubmed/SOME_ARTICLE_ID.
-
-            raise exceptions.UriResolutionError("The URI Template \"%s\" in \"%s\" is an absolute URL. "
-                                                "It should be relative and used with baseurl" %
-                                                (self.uri, pageobj_name))
-
-        if len(args) > 0:
-            # the user wants to open a non-default uri
-            uri_vars = {}
-
-            first_arg = args[0]
-            if not self._is_robot:
-                if isinstance(first_arg, basestring):
-                    # In Python, if the first argument is a string and not a dict, it's a url or path.
-                    arg_type = "url"
-                else:
-                    arg_type = "dict"
-            elif self._is_robot:
-                # We'll always get string args in Robot
-                if self._is_url_absolute(first_arg) or first_arg.startswith("/"):
-                    arg_type = "url"
-                else:
-                    arg_type = "robot"
-
-            if arg_type != "url" and hasattr(self, "uri") and uri_type == 'plain':
-                raise exceptions.UriResolutionError(
-                    "URI %s is set for page object %s. It is not a template, so no arguments are allowed." %
-                    (self.uri, pageobj_name))
-
-            if arg_type == "url":
-                if self._is_url_absolute(first_arg):
-                    # In Robot, the first argument is always a string, so we need to check if it starts with "/" or a scheme.
-                    # (We're not allowing relative paths right now._
-                    return first_arg
-                elif first_arg.startswith("//"):
-                    raise exceptions.UriResolutionError("%s is neither a URL with a scheme nor a relative path"
-                                                        % first_arg)
-                else:  # starts with "/"
-                    # so it doesn't need resolution, except for appending to baseurl and stripping leading slash
-                    # if it needed: i.e., if the base url ends with "/" and the url starts with "/".
-
-                    return re.sub("\/$", "", self.baseurl) + first_arg
-            elif arg_type == "robot":
-                # Robot args need to be parsed as "arg1=123", "arg2=foo", etc.
-                for arg in args:
-                    split_arg = arg.split("=")
-                    uri_vars[split_arg[0]] = "=".join(split_arg[1:])
-            else:
-                # dict just contains the args as keys and values
-                uri_vars = args[0]
-
-            # Check that variables are correct and match template.
-            if not self._vars_match_template(self.uri, uri_vars):
-                raise exceptions.UriResolutionError(
-                    "The variables %s do not match template %s for page object %s"
-                    % (uri_vars, self.uri, pageobj_name)
-                )
-            self.uri_vars = uri_vars
-            return uritemplate.expand(self.baseurl + self.uri, uri_vars)
-        else:
-            if uri_type == 'template':
-                raise exceptions.UriResolutionError('%s has uri template %s , but no arguments were given to resolve it' %
-                                                    (pageobj_name, self.uri))
-            # the user wants to open the default uri
-            return self.baseurl + self.uri
-
-    @staticmethod
-    def _is_url_absolute(url):
-        """
-        We're making the scheme mandatory, because webdriver can't handle just "//".
-        """
-        return re.match("^(\w+:(\d+)?)\/\/", url) is not None
 
     def log(self, msg, level="INFO", is_console=True):
         """ Logs either to Robot log file or to a file called po_log.txt
@@ -748,107 +599,11 @@ class _BaseActions(_S2LWrapper):
         self._abstracted_logger.log(msg, page_name, level, is_console)
         return self
 
-    def go_to(self, *args):
-        """
-        Wrapper to make go_to method support uri templates.
-        """
-        resolved_url = self._resolve_url(*args)
-        super(_BaseActions, self).go_to(resolved_url)
-        return self
-
-    def _generic_make_browser(self, webdriver_type, desired_cap_type, remote_url, desired_caps):
-        """Override Selenium2Library's _generic_make_browser to allow for extra params
-        to driver constructor."""
-        kwargs = {}
-        if not remote_url:
-            if 'service_args' in inspect.getargspec(webdriver_type.__init__).args:
-                kwargs['service_args'] = self.service_args
-            browser = webdriver_type(**kwargs)
-        else:
-            browser = self._create_remote_web_driver(desired_cap_type, remote_url, desired_caps)
-        return browser
-
-    def _make_browser(self, browser_name, desired_capabilities=None, profile_dir=None, remote=None):
-        creation_func = self._get_browser_creation_function(browser_name)
-
-        if not creation_func:
-            raise ValueError(browser_name + " is not a supported browser.")
-
-        browser = creation_func(remote, desired_capabilities, profile_dir)
-        browser.set_speed(self._speed_in_secs)
-        browser.set_script_timeout(self._timeout_in_secs)
-        browser.implicitly_wait(self._implicit_wait_in_secs)
-        return browser
-
-    def open(self, *args):
-        """
-        Wrapper for Selenium2Library's open_browser() that calls resolve_url for url logic and self.browser.
-        It also deletes cookies after opening the browser.
-
-        :param *args: A list or dictionary of variables mapping to a page object's uri template. For example given a
-        template like this::
-
-                class MyPageObject(PageObject):
-                    uri = "category/{category}"
-
-                    ...
-
-        calling in Python::
-
-            ...
-            my_page_object.open({"category": "home-and-garden"})
-
-        or in Robot Framework::
-
-           ...
-           Open My Page Object  category=home-and-garden
-
-        ...would open the browser at: `/category/home-and-garden`
-
-        If no `uri_var` is passed the page object tries to open the browser at its uri attribute.
-
-
-        :param delete_cookies: If set to True, deletes browser's cookies when called.
-        :type delete_cookies: Boolean
-        :returns: _BaseActions instance
-        """
-        resolved_url = self._resolve_url(*args)
-        if self._attempt_sauce:
-            remote_url = "http://%s:%s@ondemand.saucelabs.com:80/wd/hub" % (self.sauce_username, self.sauce_apikey)
-            caps = getattr(webdriver.DesiredCapabilities, self.browser.upper())
-            caps["platform"] = self.sauce_platform
-            if self.sauce_browserversion:
-                caps["version"] = self.sauce_browserversion
-            if self.sauce_device_orientation:
-                caps["device_orientation"] = self.sauce_device_orientation
-            if self.sauce_screenresolution:
-                caps["screenResolution"] = self.sauce_screenresolution
-
-            try:
-                self.open_browser(resolved_url, self.browser, remote_url=remote_url, desired_capabilities=caps)
-            except (urllib2.HTTPError, WebDriverException):
-                raise exceptions.SauceConnectionError("Unable to connect to sauce labs. Check your username and "
-                                                      "apikey")
-
-            self.session_id = self.get_current_browser().session_id
-            self.log("session ID: %s" % self.session_id)
-
-        else:
-            self.open_browser(resolved_url, self.browser)
-
-        self.set_window_size(1920, 1080)
-
-        self.log("PO_BROWSER: %s" % (str(self.get_current_browser())), is_console=False)
-
-        return self
-
-    def close(self):
-        """
-        Wrapper for Selenium2Library's close_browser.
-        :returns: None
-        """
-        self.close_browser()
-        return self
+    def wait_until_alert_is_present(self, timeout=None):
+        alert_present = False
+        self.wait_for(lambda: EC.alert_is_present(), timeout=timeout,
+                      message="No alert was present after %s"
+                              % self._format_timeout(timeout or self.selenium_implicit_wait))
 
     def wait_until_element_is_not_visible(self, locator, timeout=None, error=None):
         """Waits until element specified with `locator` is not visible.
